+++ rest_client_zrok/metadata/configuration_responses.go
@@ -30,7 +30,7 @@ func (o *ConfigurationReader) ReadResponse(response runtime.ClientResponse, cons
 		}
 		return result, nil
 	default:
-		return nil, runtime.NewAPIError("response status code does not match any response statuses defined for this endpoint in the swagger spec", response, response.Code())
+		return nil, runtime.NewAPIError("[GET /configuration] configuration", response, response.Code())
 	}
 }
 
@@ -71,6 +71,11 @@ func (o *ConfigurationOK) IsServerError() bool {
 // IsCode returns true when this configuration o k response a status code equal to that given
 func (o *ConfigurationOK) IsCode(code int) bool {
 	return code == 200
+}
+
+// Code gets the status code for the configuration o k response
+func (o *ConfigurationOK) Code() int {
+	return 200
 }
 
 func (o *ConfigurationOK) Error() string {